requirements.txt
selenium
colorama
discord.py
asyncio
dotenv